Create a session DeferredStartHandler to wait for actual session
The main purpose here is actually to populate the
MessageSession.sessionID, which in the case of NewSession(nil) isn't
known until we actually receive a message.

Beyond that I was a bit surprised to see my Start method called way
before there was actually a session present -- so this delayed start
kinda makes sense there.

On the other hand this will likely break other people's use case. :-(